Abolished Function type, introduced Funcer type
Funcers (previously Functions) are the workhorses of function call resolution.
A Funcer is a Go function which takes an input type, a function name and a
number of arguments. It tries to find a matching function. If successful, it
returns the parameters of that function, its output type and its action. It
also always returns a boolean flag indicating success or failure.

Advantages of Funcers over the old Function structs:

1. Input and output to resolution more clearly separated
2. Funcers can determine the output type dynamically based on the input type
   (first step towards generics)