Refactored Subsumes() implementations
Less abstraction, more uniformity: the body of every Subsumes() implementation
is a type switch statement now (except for AnyType, where it's just return
true). Depending on the type of the candidate subsumee, does the right thing
(hopefully). Reduced delegating to other Subsumes() implementations.

However, subsumption checking is now delegated if the candidate subsumee is a
UnionType - to the new inverseSubsumes() method. This is needed iff the
candidate subsumer is a complex non-union type. If the candidate subsumer is
a union type, we have a specialized implmentation.

Next up: implement Union() and Partition() in a similar way.