Create output shape after typechecking parameters
The order doesn't matter for RegularFuncer, but in general the output shape of
the call can depend on the output shapes of the parameters (e.g., with the `m`
funcer), so this is the more logical order. Also, if typechecking fails, we
don't need to bother to create the output shape.