Better source positions for "call tokens"
Currently, we have six kinds of tokens that are parsed into call
expressions:

1. Name (no arguments)
2. Op1Num (a 1-character operator followed by a number)
3. Op2Num (a 2-character operator followed by a number)
4. Op1Name (a 1-character operator followed by a name)
5. Op2Name (a 2-character operator followed by a name)
6. NameRegexp (a name followed by a regexp)

2-6 have 1 argument, which is parsed into a separate Expression object
in the AST. The source position off that Expression used to be the same
as that of the call token. Now its column is properly offset by the
length of the leading operator/name.

Also, stylistically, made NameRegexp's Ast method more similar to the
others'.