Typechecker now keeps track of IDs
In addition to output shapes and actions, typechecking an expression now also
returns a list of IDs of variables and parameters referenced in the expression.

The plan is to use this information to reduce input variable stacks to what is
actually needed, in order to plug that memory leak in tail recursion.