Merged Map, Obj into a more powerful Obj type
Obj types now specify *rest types* (default Any), a type that the values of all
keys not mentioned in the object type inhabit. So, e.g., Map<Str> is now
Obj<Str>, and we can also mix it: Obj<start: Num, Str> - objects with a numeric
start value and any other values are Str. Obj<a: Num> is now short for Obj<a:
Num, Any>. This gives us back the old "open-world assumption", allowing us to
use the type system for checking object pattern matching again. "Closed-world"
object types are still possible, just specify Void as the rest type: Obj<a:
Num, b: Str, Void> has a numeric value for a and a string value for b *and
nothing else*.