Predicate syntax
Conditionals now support an additional syntax where there are no `then`
clauses, no `else` clause and no final `ok`. For an input value x, all positive
branches return `{yes: x}`, and the negative branch returns `{no: x}`. This
"predicate form" of the conditional is intended to be used in arguments of
"higher-order" funcers such as keepWhile. The lack of the final `ok` in fact
would make the form impractical to use anywhere but funcer arguments.